Manage duplicate events
Why: It's possible for the same event to appear several times on the
calendar. If we fetch the same event multiple times, that is a waste of
resources.

* ...

Converted the links array into a hash map. so if the event id is already
in the hash map, just overwrite with the same link, lazy de-dupe

* ...