Setup for heroku deployment
Why: found the Heroku allows for backgroun worker process. This can be configured through the Procfile

*---

How: Created a procfile that defines a worker thread. This worker will run the fetcher process as a separate process. Adjusted tsconfig to generate compiled JS next to typescript files. This allows the module paths to be relatively the same

*---